Sharpen 2D Simplex and remove seams.
This makes the noise look worse when zoomed out and only one octave is used, but a little better when multiple octaves are used, and a lot better when zoomed in very far. I tried a constant computed by kdotjpg for a different (though similar) set of gradients, and it also had seams, even when I used his gradients. Shrinking the range very slightly worked.